feat(admin): refactor admin page
