+++ src/Root.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import PageLoading from './components/loading/loading'
 import RouterContainer from './components/router-container/router'
 import NotFound from './pages/not-found/not-found'
-import AdminPanel from './pages/admin/admin'
+import AdminPanel from './pages/dash/[userid]/admin'
 import IndexPage from './pages/index/index'
 import GithubOAuthPage from './pages/auth/github'
 import AuthPage from './pages/auth/auth'
