--- src/components/card/card.css
@@ -1,3 +1,0 @@
-.card {
-  /* backdrop-filter: blur(10px); */
-}
